requirements.txt
aws-cdk.core
aws-cdk.aws_ec2
aws-cdk.aws_ecs
aws-cdk.aws_efs
aws-cdk.aws_cloudfront
aws-cdk.aws_cloudfront_origins
aws-cdk.aws_ecs_patterns
aws-cdk.aws-autoscaling-common
